minor bug fix 4

+++ contact.c
@@ -125,21 +125,21 @@ void createContact(AddressBook *addressBook)
         }
 
         if(flag1 == 0 && strcmp(addressBook->contacts[index_name[i]].name, name) > 0) {
-            for(int k = addressBook->contactCount - 1; k >= i; k--) {
+            for(int k = (int)addressBook->contactCount - 1; k >= (int)i; k--) {
                 index_name[k + 1] = index_name[k];
             }
             index_name[i] = addressBook->contactCount;
             flag1 = 1;
         }
         if(flag2 == 0 && strcmp(addressBook->contacts[index_phone[i]].phone, phone) > 0) {
-            for(int k = addressBook->contactCount - 1; k >= i; k--) {
+            for(int k = (int)addressBook->contactCount - 1; k >= (int)i; k--) {
                 index_phone[k + 1] = index_phone[k];
             }
             index_phone[i] = addressBook->contactCount;
             flag2 = 1;
         }
         if(flag3 == 0 && strcmp(addressBook->contacts[index_email[i]].email, email) > 0) {
-            for(int k = addressBook->contactCount - 1; k >= i; k--) {
+            for(int k = (int)addressBook->contactCount - 1; k >= (int)i; k--) {
                 index_email[k + 1] = index_email[k];
             }
             index_email[i] = addressBook->contactCount;
